requirements.txt
# requirements.txt
requests
influxdb
python-dateutil
pytz
